Added logic to fill Tasks widget
This resulted in a minor loss of efficiency (in the range of 1% overall
CPU usage or so)

We'll eventually need a broader refactor of the way we query for system
processes to improve efficiency. This will likely involve replacing the
Processes.processes (a dict at the moment) with a list of namedtuples or
something like that so we're not spending resources on hashes anymore.